04_Deployement/producer/requirements.txt
numpy<2
pandas==2.1.4
python-dotenv==1.0.1
requests==2.31.0
confluent-kafka==2.3.0
